requirements.txt
tqdm
numpy
aiohttp
beautifulsoup4
requests
io
tk
pywebview